api timeout fix

+++ chess/ChessGame.cpp
@@ -12,23 +12,26 @@ ChessGame::ChessGame(int scale, string name, GameMode gameMode): ChessWindow(sca
     cout << "game name: " << name << endl;
     this->scale = scale;
     this->name = name;
+    this->gameMode = gameMode;
     this->notificationLabel = new Label("", 20, this);
     this->board = new ChessBoard(scale, gameMode, this);
     this->pickedUpPiece = nullptr;
     this->pickedUpPieceTile = nullptr;
     
-    drawButton = new Button(0.8, 0.4, 0.15, 0.075, this, [&] {cout << "Ask for draw" << endl;});
-    drawButton->setColor(Color(96, 108, 131));
-    drawButton->setLabel("Ask For Draw");
-    
-    secedeButton = new Button(0.8, 0.6, 0.15, 0.075, this, [&] {
-        ChessServerAPI::deleteGame(board, [] {});
-        this->renderWindow->close();
-        MainWindow* mainWindow = new MainWindow();
-        mainWindow->launch();
-    });
-    secedeButton->setColor(Color(96, 108, 131));
-    secedeButton->setLabel("Surrender");
+    if (gameMode == MULTI_PLAYER) {
+        drawButton = new Button(0.8, 0.4, 0.15, 0.075, this, [&] {cout << "Ask for draw" << endl;});
+        drawButton->setColor(Color(96, 108, 131));
+        drawButton->setLabel("Ask For Draw");
+        
+        secedeButton = new Button(0.8, 0.6, 0.15, 0.075, this, [&] {
+            ChessServerAPI::deleteGame(board, [] {});
+            this->renderWindow->close();
+            MainWindow* mainWindow = new MainWindow();
+            mainWindow->launch();
+        });
+        secedeButton->setColor(Color(96, 108, 131));
+        secedeButton->setLabel("Surrender");
+    }
     
     this->board->setScale(scale);
     
@@ -40,6 +43,15 @@ void ChessGame::resizeEvent(Event event) {
     this->update();
     this->drawButton->update();
     this->secedeButton->update();
+}
+
+void ChessGame::closeEvent(Event event) {
+    if (gameMode == MULTI_PLAYER) {
+        ChessServerAPI::deleteGame(board, [] {});
+    }
+    this->renderWindow->close();
+    MainWindow* mainWindow = new MainWindow();
+    mainWindow->launch();
 }
 
 void ChessGame::update() {
